brobot199/requirements.txt
git+https://github.com/Rapptz/discord.py.git
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
pymongo==3.10.1
requests==2.24.0
pillow==7.2.0